requirements.txt
langchain
langgraph
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
langchain-openai
tavily-python
langchain-community
mcp
